Refactor for Capybara
